+++ src/test/java/org/greencodeinitiative/creedengo/python/checks/AvoidCreatingTensorUsingNumpyOrNativePythonTest.java
@@ -17,15 +17,14 @@
  */
 package org.greencodeinitiative.creedengo.python.checks;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.sonar.python.checks.utils.PythonCheckVerifier;
 
-
-
 public class AvoidCreatingTensorUsingNumpyOrNativePythonTest {
+
     @Test
     public void test() {
-      PythonCheckVerifier.verify("src/test/resources/checks/avoidCreatingTensorUsingNumpyOrNativePython.py", new AvoidCreatingTensorUsingNumpyOrNativePython());
+        PythonCheckVerifier.verify("src/test/resources/checks/avoidCreatingTensorUsingNumpyOrNativePython.py", new AvoidCreatingTensorUsingNumpyOrNativePython());
     }
     
 }